+++ app/controllers/components_controller.rb
@@ -5,7 +5,7 @@ class ComponentsController < ApplicationController
 
   def create
     page = Page.find(params[:page_id])
-    authorize_wizard(page.wizard)
+    authorize page, :update?
     form = AddComponentForm.new(page: page, kind: params.require(:component)[:kind])
     form.save
 
@@ -14,7 +14,7 @@ class ComponentsController < ApplicationController
 
   def update
     component = Component.find(params[:id])
-    authorize_wizard(component.page.wizard)
+    authorize component, :update?
     form = ComponentForm.from_component(component)
     form.assign_attributes(component_params)
     form.save
@@ -27,7 +27,7 @@ class ComponentsController < ApplicationController
 
   def destroy
     component = Component.find(params[:id])
-    authorize_wizard(component.page.wizard)
+    authorize component, :destroy?
     component.destroy!
 
     respond_with_panes(component.page, builder: true)
